networkd: fix UAF and leaks in config_parse_dhcp_static_leases
Add the fuzz test file that triggered the issues.

A UAF appeared in config_parse_dhcp_static leases as DHCPClientIds were
freed but not removed from the static leases hashmap.
When the function was called multiple times, the entries would be
accessed when calling ordered_hashmap_replace which caused a UAF.
Additionally, after the DHCPClientId c was freed, c->data wasn't so a
custom free function is added here.

The function also leaked the lease, as TAKE_PTR nulled the pointer so it
wouldn't be automatically freed.

Signed-off-by: Juraj Vijtiuk <juraj.vijtiuk@sartura.hr>